# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "The Digital Cellar BE API",
        "version": "1"
    },
    "paths": {
        "/api/v1/auth/sign-in": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "operationId": "postSignInUser",
                "requestBody": {
                    "request": "UserStoreRequestBody",
                    "description": "Post new user object to the system",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "user email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "user password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "oqh12894nnlas******"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/sign-out": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "operationId": "postSignOutUser",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get Simple Pagination Laravel of data Users",
                "description": "Returns list of simple pagination Laravel Users",
                "operationId": "getUsers",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Target page, default value is one",
                        "required": false
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "per_page, default value is ten",
                        "required": false
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "global filter users, with name and email",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserCollection"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Post new user",
                "description": "Post new user with name and email, created_at and updated_at is nullable",
                "operationId": "postUser",
                "requestBody": {
                    "request": "UserStoreRequestBody",
                    "description": "Post new user object to the system",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/UserResource"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "nullable": "true",
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/UserResource"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/{id_user}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user with partial user data",
                "description": "Get with partial user data",
                "operationId": "getUser",
                "parameters": [
                    {
                        "name": "id_user",
                        "in": "path",
                        "description": "Target user id",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "nullable": "true",
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/UserResource"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete user",
                "description": "Permanently delete user",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "id_user",
                        "in": "path",
                        "description": "Target user id",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "boolean",
                                            "example": true,
                                            "nullable": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Patch user with partial user data",
                "description": "Patch with partial user data",
                "operationId": "patchUser",
                "parameters": [
                    {
                        "name": "id_user",
                        "in": "path",
                        "description": "Target user id",
                        "required": false
                    }
                ],
                "requestBody": {
                    "request": "UserUpdateRequestBody",
                    "description": "Patch user object to the system",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/UserResource"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "nullable": "true",
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/UserResource"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "UserCollection": {
                "properties": {
                    "data": {
                        "description": "User details",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserResource"
                        },
                        "nullable": false
                    },
                    "links": {
                        "description": "User details",
                        "properties": {
                            "links": {
                                "type": "string",
                                "example": "http://127.0.0.1:8000/api/v1/users?page=1",
                                "nullable": false
                            },
                            "last": {
                                "type": "string",
                                "example": null,
                                "nullable": false
                            },
                            "next": {
                                "type": "string",
                                "example": null,
                                "nullable": false
                            },
                            "prev": {
                                "type": "string",
                                "example": null,
                                "nullable": false
                            }
                        },
                        "type": "object",
                        "nullable": false
                    },
                    "meta": {
                        "description": "User details",
                        "properties": {
                            "links": {
                                "type": "string",
                                "example": "http://127.0.0.1:8000/api/v1/users?page=1",
                                "nullable": false
                            },
                            "current_page": {
                                "type": "number",
                                "example": 1,
                                "nullable": false
                            },
                            "from": {
                                "type": "number",
                                "example": 1,
                                "nullable": false
                            },
                            "path": {
                                "type": "string",
                                "example": "http://127.0.0.1:8000/api/v1/users",
                                "nullable": false
                            },
                            "per_page": {
                                "type": "number",
                                "example": 10,
                                "nullable": false
                            },
                            "to": {
                                "type": "number",
                                "example": 2,
                                "nullable": false
                            }
                        },
                        "type": "object",
                        "nullable": false
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "properties": {
                    "id": {
                        "description": "user unique id / primary key",
                        "type": "number",
                        "example": 1,
                        "nullable": "true"
                    },
                    "name": {
                        "description": "User name",
                        "type": "string",
                        "example": "Kellen Boyer",
                        "nullable": "false"
                    },
                    "email": {
                        "description": "User E-mail",
                        "type": "string",
                        "example": "kellen.boyer@example.com",
                        "nullable": "false"
                    },
                    "password": {
                        "description": "User password, will be hashed",
                        "type": "string",
                        "example": "password",
                        "nullable": "false"
                    },
                    "created_at": {
                        "description": "User created at",
                        "type": "string",
                        "example": "2023-11-24T01:10:37.000000Z",
                        "nullable": "true"
                    },
                    "updated_at": {
                        "description": "User updated at",
                        "type": "string",
                        "example": "2023-11-24T01:10:37.000000Z",
                        "nullable": "true"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}